Extract project creation form into its own component
